+++ src/commands/hardis/work/save.ts
@@ -6,6 +6,7 @@ import * as c from 'chalk';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
+import { ResetMode } from 'simple-git';
 import { execCommand, execSfdxJson, getCurrentGitBranch, git, gitHasLocalUpdates, interactiveGitAdd, uxLog } from '../../../common/utils';
 import { prompts } from '../../../common/utils/prompts';
 import { getConfig, setConfig } from '../../../config';
@@ -55,6 +56,11 @@ export default class SaveTask extends SfdxCommand {
     const localBranch = await getCurrentGitBranch();
 
     uxLog(this, c.cyan(`This script will prepare the merge request from your local branch ${c.green(localBranch)} to remote ${c.green(config.developmentBranch)}`));
+
+    // Unstage files
+    await git().reset(ResetMode.SOFT);
+
+    // Pull from scratch org
     uxLog(this, c.cyan(`Pulling sources from scratch org ${this.org.getUsername()}...`));
     const pullCommand = 'sfdx force:source:pull -w 60 --forceoverwrite';
     await execCommand(pullCommand, this, { output: true, fail: true });