+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- Define VsCode as git merge/diff tool if none is defined
+- Unstash changes (git reset) at the beginning of hardis:work:save
+
 ## [2.7.2] 2021-03-30
 
 - Check user is sure to want to reuse an existing scratch org